🚚 Rename: fetch 함수 폴더 이동(apis → apis/base)
