+++ src/shared/apis/base/publicFetch.ts
@@ -3,7 +3,7 @@ import {
   FetchRequestOptions,
   QueryParams,
   RequestConfig,
-} from '@/shared/apis/coreFetch';
+} from '@/shared/apis/base/coreFetch';
 import { ENV } from '@/shared/apis/env';
 /**
  * 토큰 없는 공개 API용 Fetch 유틸리티
